+++ src/main/java/com/crisalis/crisalisback/dto/ProductoDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @Builder
@@ -11,5 +13,5 @@ public class ProductoDTO {
     private String nombre;
     private String descripcion;
     private int stock;
-    private double precioBase;
+    private BigDecimal precioBase;
 }
