+++ src/main/java/com/crisalis/crisalisback/dto/ServicioDTO.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @Builder
 public class ServicioDTO {
     private String nombre;
     private String descripcion;
-    private double precioBase;
-    private double precioSoporte;
+    private BigDecimal precioBase;
+    private BigDecimal precioSoporte;
 }
