+++ src/main/java/com/crisalis/crisalisback/model/Descuento.java
@@ -13,6 +13,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,7 +24,7 @@ public class Descuento {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
 
-    private double descuentoGenerado;
+    private BigDecimal descuentoGenerado;
 
     @OneToOne
     @MapsId
