+++ src/main/java/com/crisalis/crisalisback/model/Servicio.java
@@ -8,6 +8,7 @@ import com.crisalis.crisalisback.dto.ProductoDTO;
 import com.crisalis.crisalisback.dto.ServicioDTO;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,14 +18,14 @@ import java.util.List;
 @Entity
 @DiscriminatorValue("servicio")
 public class Servicio extends ProductoBase{
-    private double precioSoporte;
+    private BigDecimal precioSoporte;
 
     /*@Getter(value = AccessLevel.NONE)
     @Setter(value = AccessLevel.NONE)
     @OneToMany(mappedBy = "servicio")
     private List<ServicioPedido> listaServiciosPedidos = new ArrayList<ServicioPedido>();*/
 
-    public Servicio(String tipo, String nombre, String descripcion, double precioBase, double precioSoporte) {
+    public Servicio(String tipo, String nombre, String descripcion, BigDecimal precioBase, BigDecimal precioSoporte) {
         super(tipo, nombre, descripcion, precioBase);
         this.precioSoporte = precioSoporte;
     }
